move deprecation wrapper inside qip folder
