+++ qutip/__init__.py
@@ -226,58 +226,11 @@ from qutip.interpolate import *
 from qutip.scattering import *
 
 ########################################################################
+# This section exists only for the deprecation warning of qip importation.
+# It can be deleted for a major release.
+
 # quantum information
-# from qutip.qip import *
-
-# This section exists only for the deprecation of
-# importing members of the qip submodule into the qutip namespace.
-
-import inspect
-from functools import wraps as _func_wrap
-import qutip.qip as _temp_qip
-from qutip.qip_depracation import _qip_importation_warning
-
-
-def _qip_func_wrapper(func):
-    """Function wrapper for adding a deprecation warning."""
-    @_func_wrap(func)
-    def deprecated_func(*args, **kwargs):
-        _qip_importation_warning()
-        return func(*args, **kwargs)
-    return deprecated_func
-
-
-#  Wrap all qip functions with a deprecation warning
-_func_pairs = inspect.getmembers(_temp_qip, inspect.isfunction)
-for _name, _func in _func_pairs:
-    locals()[_name] = _qip_func_wrapper(_func)
-del _name, _func, _func_pairs, _qip_func_wrapper
-
-
-def _qip_class_wrapper(original_cls):
-    """Class wrapper for adding a deprecation warning."""
-    class Deprecated_cls(original_cls):
-        def __init__(self, *args, **kwargs):
-            _qip_importation_warning()
-            super(Deprecated_cls, self).__init__(*args, **kwargs)
-    # copy information form the original class, similar to functools.wraps
-    for attr in ('__module__', '__name__', '__qualname__', '__doc__'):
-        try:
-            value = getattr(original_cls, attr)
-        except AttributeError:
-            pass
-        else:
-            setattr(Deprecated_cls, attr, value)
-    return Deprecated_cls
-
-
-#  Wrap all qip classes with a deprecation warning
-_cls_pairs = inspect.getmembers(_temp_qip, inspect.isclass)
-for _name, _cls in _cls_pairs:
-    locals()[_name] = _qip_class_wrapper(_cls)
-del _name, _cls, _cls_pairs, _qip_class_wrapper
-
-del _temp_qip
+from qutip.qip import *
 ########################################################################
 
 # utilities
