+++ qutip/qip/__init__.py
@@ -30,9 +30,8 @@
 #    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 #    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ###############################################################################
-import qutip.qip.ops
-import qutip.qip.device
-from qutip.qip.ops import *
-from qutip.qip.circuit import *
-from qutip.qip.qubits import *
-from qutip.qip.device import *
+# This is used for the deprecation.
+# To remove the warning, and finalize the deprecation
+# simply delete the file qutip.qip.qip_depracation.py
+# and the line bellow.
+from qutip.qip.qip_depracation import *
